Move hisdisplay to hisdisplaycomponent
